Users flashcard is updated to the appstate #1
