+++ src/index.js
@@ -1,4 +1,4 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import Card from './card'
-ReactDOM.render(<Card />, document.getElementById('root'))
+import App from './app'
+ReactDOM.render(<App />, document.getElementById('root'))
